Set RestartLight in init instead of first cycle
The RestartLight is currently set during the first cycle of the emulator
(CycleCounter). This presents an issue in NASSP in case the emulator is
resumed (We don't restore CycleCounter due to an issue with the
CduFifos) where the restart light is always turned on at simulation
start.

Doing this in agc_engine_init instead keeps the same behaviour but
allows NASSP to turn it off in the first cycle in case of a resumed simulation.